// frontend/src/api/model/outStudentOut.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * TT API
 * Team tuner API documentation
 * OpenAPI spec version: 1.0
 */

export interface OutStudentOut {
  agreeableness?: number;
  completedQuestioner: boolean;
  completedQuestionerAt?: string;
  concientiousness?: number;
  createdAt: string;
  extraversion?: number;
  githubUsername: string;
  id: string;
  neuroticism?: number;
  openness?: number;
  univeresityId?: string;
  updatedAt: string;
}